Refactor: jtb114 - 8/28/25 - 1:09pm ----
On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
      	new file:   internal/files/input_file.go
	new file:   internal/files/vcf_streamer.go
	new file:   internal/main.go

Changes not staged for commit:
	modified:   cmd/find_individuals_variants.go
	modified:   cmd/pull_variants.go
	modified:   go.mod
	modified:   scripts/pull_variants_wrapper.sh

Abstracted some of the input file handling:
---
The filehandling was getting very verbose. Therefore code was refactored so to Filereader that can handle both compressed files and non compressed files and then a vcf streamer that can handle input streams. This change removes the responsiblity for creating opening all the filehandles, creating the scanner, updating the scanner's buffer, closing the file, etc... from the main commands.