Feature: jtb114 - 08/26/25 - 1:14pm ---- On branch main Your branch is ahead of 'origin/main' by 1 commit.    (use "git push" to publish your local commits)
Changes to be committed:
	modified:   scripts/pull_variants_wrapper.sh

Changes not staged for commit:
	modified:   cmd/find_individuals_variants.go
	modified:   cmd/pull_variants.go
	modified:   go.mod

Added a runner script:
---
This tool is designed to be run in combination with bcftools. I created a bash script that mergs the different functionality so that the user can just run the wrapper instead of having to manually chain together the commands.