Fix: jtb114 - 8/26/25 - 3:51pm --- On branch main Your branch is ahead of 'origin/main' by 3 commits.    (use "git push" to publish your local commits)
Changes to be committed:
	modified:   internal/files/input_file.go
	new file:   pull_variants_wrapper.sh

Fixed the wrapper script so that the help message properly displays. Still need to update the script so that it accounts for when no arguments were provided. Additionally added an argument for the consequence column since we are also filtering on that in the file. Finally made the close statement for the abstracted fileReader more idiomatic