Fix: - jtb114 - 8/27/25 - 11:09am ---
On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	modified:   cmd/find_individuals_variants.go
	modified:   internal/files/input_file.go

fixed defered file closing and fixed a bug when value was not provided
with the consequence col tag
---
The filehandles were prematurely closing anytime the FileReader object was used even with the defered function. This closure is likely due to to one of the underlying readers from os.Open or GzipReader closing early and then go panics when the bufio.Scanner attempts to read from the file. I moved the deferred closing into an anynonymous function that is underneath wherever the fileReader gets created (There may be a better solution but at the time of writing, I don't know it). There was also a bug where if the user passed the '--consequence-col' flag to the view-sample-variants command but didn't provide a value then the dictionary lookup would return an error but the program would hide the error. I made it explicity that the error happened and then made the program terminate early because the results are not reliable at that point.