Feature: - jtb114 - 8/27/25 - 2:19pm --- On branch main Your branch is up to date with 'origin/main'.
Changes to be committed:
	modified:   cmd/find_variant_calls.go
	modified:   cmd/pull_variants.go
	modified:   internal/files/input_file.go
	deleted:    internal/files/vcf_streamer.go

generalized the fileReader struct and added new functionality to pull all the carriers for a variant of interest
----
Instead of having a separate vcf_streamer struct I created a new VCFReader struct in the input_file.go. This struct is composed of the FileReader so it can keep all the attributes of that struct but it uses a modified ParseHeader function that creates not only the header column mapping of a map[string]int but also a sample mapping of map[int]string. This change needed to be made because when streaming the vcf file it is easy to miss which samples are which and then you cannot correctly print out the information.

Also created another function called 'find-all-carriers' that takens input streamed in from BCFtools for a reigon of interest (preferably a single site) and returns a file telling who is variant at that site.